+++ analysis/layout/src/lib.rs
@@ -2,24 +2,24 @@ use support::prelude::*;
 
 const MAX: usize = 30_000;
 
-fn test_direct<const K: usize>(grow: f64, val: [u8; K]) -> Vec<(usize, f64, f64)> {
+fn test_direct<const K: usize>(grow: f64, val: [u8; K]) -> Vec<(i128, f64, f64)> {
     let mut all: Vec<_> = dist::dist(grow).map(|x| Dummy(x, val)).take(MAX).collect();
     test_one_kind(&mut all)
 }
 
-fn test_indirect<const K: usize>(grow: f64, val: [u8; K]) -> Vec<(usize, f64, f64)> {
+fn test_indirect<const K: usize>(grow: f64, val: [u8; K]) -> Vec<(i128, f64, f64)> {
     let mut all: Vec<_> = dist::dist(grow).map(|x| Dummy(x, val)).take(MAX).collect();
     let mut all: Vec<_> = all.iter_mut().collect();
     test_one_kind(&mut all)
 }
 
-fn test_default<const K: usize>(grow: f64, val: [u8; K]) -> Vec<(usize, f64, f64)> {
+fn test_default<const K: usize>(grow: f64, val: [u8; K]) -> Vec<(i128, f64, f64)> {
     let mut all: Vec<_> = dist::dist(grow).map(|x| Dummy(x, val)).take(MAX).collect();
     let mut all: Vec<_> = all.iter_mut().map(|x| Dummy(x.0, &mut x.1)).collect();
     test_one_kind(&mut all)
 }
 
-fn test_one_kind<T: ColfindHandler>(all: &mut [T]) -> Vec<(usize, f64, f64)> {
+fn test_one_kind<T: ColfindHandler>(all: &mut [T]) -> Vec<(i128, f64, f64)> {
     let mut plots = vec![];
     for a in n_iter(0, MAX) {
         let bots = &mut all[0..a];
@@ -31,7 +31,7 @@ fn test_one_kind<T: ColfindHandler>(all: &mut [T]) -> Vec<(usize, f64, f64)> {
             tree
         });
 
-        plots.push((a, construct_time, query_time));
+        plots.push((a as i128, construct_time, query_time));
     }
     plots
 }
@@ -43,7 +43,7 @@ pub enum Layout {
 }
 
 #[inline(never)]
-pub fn bench(typ: Layout, grow: f64, size: usize) -> Vec<(usize, f64, f64)> {
+pub fn bench(typ: Layout, grow: f64, size: usize) -> Vec<(i128, f64, f64)> {
     match typ {
         Layout::Direct => match size {
             8 => test_direct(grow, [0u8; 8]),
